Correct lookup table provisioning

+++ modules/data_warehouse/src/sql/sp_provision_lookup_tables.sql
@@ -16,8 +16,8 @@ CREATE OR REPLACE TABLE `${project_id}.ds_edw.distribution_centers`
     (
         id INTEGER,
         name STRING,
-        latitude FLOAT,
-        longitude FLOAT,
+        latitude FLOAT64,
+        longitude FLOAT64,
         distribution_center_geom GEOGRAPHY
     )
 AS
